+++ common/src/main/java/net/spacerulerwill/skygrid_reloaded/util/CreateWorldScreenExtension.java
@@ -1,0 +1,10 @@
+package net.spacerulerwill.skygrid_reloaded.util;
+
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.level.dimension.LevelStem;
+
+import java.util.Map;
+
+public interface CreateWorldScreenExtension {
+    Map<ResourceKey<LevelStem>, LevelStem> skygrid_reloaded$getDefaultLevelStems();
+}