+++ forge/CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added support for modded dimensions.
+
 # 1.2.2-1.21.5-forge - 25/03/25
 
 * Ported from **1.2.2-1.21.5-fabric** to **1.2.2-1.21.5-forge**